Move isomorphism implementation to homomorphism.py
